update: shared component how: put it in directory which has same name
