+++ lib/component/local/attendance_modal/attendance_modal.dart
@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../shared/navigator.dart';
-import '../../shared/riverpod.dart';
-import '../../shared/single/buttons.dart';
-import '../../shared/single/color.dart';
-import '../../shared/single/custom_icon_button.dart';
-import '../../shared/single/department_tag.dart';
-import '../../shared/single/space_box.dart';
-import '../../shared/single/text_style.dart';
+import '../../shared/single/navigator/navigator.dart';
+import '../../shared/single/riverpod/riverpod.dart';
+import '../../shared/single/buttons/buttons.dart';
+import '../../shared/single/color/color.dart';
+import '../../shared/single/department_tag/custom_icon_button.dart';
+import '../../shared/single/department_tag/department_tag.dart';
+import '../../shared/single/space_box/space_box.dart';
+import '../../shared/single/text_style/text_style.dart';
 import '../attendance_modal_with_reason/attendance_modal_with_reason.dart';
 import '/domain/class/class.dart';
 import '/domain/student/student.dart';
