+++ lib/component/local/attendance_modal_with_reason/elements/select_class_checkbox/elements/checkbox_with_class.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../../shared/single/color.dart';
-import '../../../../../shared/single/space_box.dart';
-import '../../../../../shared/single/text_style.dart';
+import '../../../../../shared/single/color/color.dart';
+import '../../../../../shared/single/space_box/space_box.dart';
+import '../../../../../shared/single/text_style/text_style.dart';
 
 class CheckboxWithClass extends StatelessWidget {
   const CheckboxWithClass({
