+++ lib/component/local/grades_table/grades_table_body/elements/curriculum_subject/curriculum_subject_body.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../../shared/single/color.dart';
-import '../../../../../shared/single/text_style.dart';
+import '../../../../../shared/single/color/color.dart';
+import '../../../../../shared/single/text_style/text_style.dart';
 import 'elements/curriculum_subject_card.dart';
 
 List curriculumSubjectList = [
