+++ lib/component/local/grades_table/grades_table_body/elements/grades_table_shared/gradation_container.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../shared/single/color.dart';
+import '../../../../../shared/single/color/color.dart';
 
 class GradationContainer extends StatelessWidget {
   GradationContainer({
