+++ lib/component/local/grades_table/grades_table_body/elements/grades_table_shared/grades_table_elements_card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../../shared/single/color.dart';
-import '../../../../../shared/single/text_style.dart';
+import '../../../../../shared/single/color/color.dart';
+import '../../../../../shared/single/text_style/text_style.dart';
 
 class GradesTableElementsCard extends StatelessWidget {
   GradesTableElementsCard({super.key, required this.elementText});
