+++ lib/component/local/grades_table/grades_table_body/elements/todays_lesson/elements/elements/sized_box_with_central_border.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../../../shared/single/color.dart';
+import '../../../../../../../shared/single/color/color.dart';
 
 class SizedBoxWithCentralBorder extends StatelessWidget {
   SizedBoxWithCentralBorder({
