+++ lib/component/local/grades_table/grades_table_body/elements/todays_lesson/elements/todays_lesson_header.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../../../shared/single/space_box.dart';
-import '../../../../../../shared/single/color.dart';
-import '../../../../../../shared/single/text_style.dart';
+import '../../../../../../shared/single/space_box/space_box.dart';
+import '../../../../../../shared/single/color/color.dart';
+import '../../../../../../shared/single/text_style/text_style.dart';
 import '../../../../grades_table_modal/grades_table_modal_show_todays_lessons.dart';
 
 class TodaysLessonHeader extends StatelessWidget {
