+++ lib/component/local/grades_table/grades_table_body/elements/todays_lesson/elements/todays_lesson_tab.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../../../shared/single/color.dart';
+import '../../../../../../shared/single/color/color.dart';
 
 class TodaysLessonTab extends StatelessWidget {
   TodaysLessonTab({super.key, required this.todaysLessonTabController});
