+++ lib/component/local/grades_table/grades_table_modal/grades_table_modal_show_todays_lessons.dart
@@ -3,7 +3,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:vantan_connect/component/local/grades_table/grades_table_body/elements/todays_lesson/elements/todays_lesson_tab_bar_view.dart';
 
 import '../../../../page/grades_table_page.dart';
-import '../../../shared/single/color.dart';
+import '../../../shared/single/color/color.dart';
 import '../grades_table_body/elements/todays_lesson/elements/todays_lesson_header.dart';
 import '../grades_table_body/elements/todays_lesson/elements/todays_lesson_tab.dart';
