+++ lib/component/local/timetable_content/element/timetable_content_case_of_basic_a.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../shared/single/space_box.dart';
+import '../../../shared/single/space_box/space_box.dart';
 import '/component/local/lesson_detail_body/hooks.dart';
 import '/domain/class/class.dart';
 import 'element/lesson.dart';
