+++ lib/component/local/timetable_header/timetable_header.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../shared/single/color.dart';
-import '../../shared/single/space_box.dart';
-import '../../shared/single/text_style.dart';
+import '../../shared/single/color/color.dart';
+import '../../shared/single/space_box/space_box.dart';
+import '../../shared/single/text_style/text_style.dart';
 
 class TimetableHeader extends StatelessWidget {
   const TimetableHeader({Key? key}) : super(key: key);
