+++ lib/component/shared/single/original_toggle_buttons/original_toggle_buttons.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../shared/single/space_box.dart';
+import '../space_box/space_box.dart';
 
 class OriginalToggleIconText extends StatelessWidget {
   const OriginalToggleIconText({
