+++ lib/component/shared/single/registerd_lesson_by_student_mark/registerd_lesson_by_student_mark.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../shared/single/color.dart';
-import 'text_style.dart';
+import '../color/color.dart';
+import '../text_style/text_style.dart';
 
 class RegisteredLessonByStudentMark extends StatelessWidget {
   const RegisteredLessonByStudentMark({Key? key}) : super(key: key);
