+++ lib/component/shared/single/riverpod/riverpod.dart
@@ -7,7 +7,7 @@ import '/domain/student/student.dart';
 import '/infrastructure/student_repository.dart';
 import '/infrastructure/query_service_repository.dart';
 import '/use_case/actor/student.dart';
-import '../../use_case/query_service/query_service.dart';
+import '../../../../use_case/query_service/query_service.dart';
 
 final studentUseCase = Provider((ref) => StudentUseCase(StudentRepository()));
