+++ lib/component/shared/single/text_style/text_style.dart
