+++ lib/page/student_list_lesson_detail_page.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:vantan_connect/component/local/attendance_record/attendance_record.dart';
-import '../component/shared/single/color.dart';
-import '../component/shared/single/text_style.dart';
+import '../component/shared/single/color/color.dart';
+import '../component/shared/single/text_style/text_style.dart';
 import '../domain/class/class.dart';
 
 class StudentListInLessonDetailPage extends ConsumerWidget {
